fix: PostgreSQL reserved word and type casting errors in migrations
## Problem
Two migration files had SQL errors preventing deployment:
1. `20251112000003_create_spatial_functions.sql` - Reserved word conflict
2. `20251112000005_create_rate_limiting_tables.sql` - Type mismatch in advisory lock

## Root Causes
1. Column name `trim` is a PostgreSQL reserved word (built-in string function)
   - Error: "syntax error at or near 'trim'" on line 28
   - Fix: Escaped as `"trim"` in RETURNS TABLE and SELECT statements

2. Type mismatch in get_filter_options_by_location function
   - Error: "Returned type VARCHAR[] does not match expected type TEXT[]"
   - Fix: Added explicit `::TEXT` casting for make, body_style, and condition columns

3. Advisory lock function signature error
   - Error: "function pg_advisory_xact_lock(integer, bigint) does not exist"
   - Fix: Changed second argument from `::BIGINT` to `::INTEGER`

## Changes Made

### File: supabase/migrations/20251112000003_create_spatial_functions.sql
- Line 28: `trim VARCHAR(100)` → `"trim" VARCHAR(100)`
- Line 63: `v.trim` → `v."trim"`
- Lines 166-168: Added `::TEXT` casting to array aggregations

### File: supabase/migrations/20251112000005_create_rate_limiting_tables.sql
- Line 60: Changed `EXTRACT(EPOCH FROM v_window_start)::BIGINT` to `::INTEGER`

### New File: supabase/migrations/20251112000006_fix_function_types.sql
- Standalone migration to update existing functions with fixes
- Can be applied via Supabase dashboard SQL Editor

## Testing
- [x] Migration files updated with fixes
- [x] Functions can be created without syntax errors
- [ ] Run SQL via Supabase dashboard to apply fixes to existing deployment

## Related Issues
Resolves 500 Internal Server Error on /api/filter-options endpoint

+++ supabase/migrations/20251112000005_create_rate_limiting_tables.sql
@@ -57,7 +57,7 @@ BEGIN
   -- Combine hash with epoch to reduce collision risk under high load
   PERFORM pg_advisory_xact_lock(
     hashtext(p_identifier || '::' || p_endpoint || '::' || v_window_start::TEXT),
-    EXTRACT(EPOCH FROM v_window_start)::BIGINT
+    EXTRACT(EPOCH FROM v_window_start)::INTEGER
   );
 
   -- Get or create counter using upsert
